refactor: remove unnecessary build_core folder
