+++ example/globalConfig.json
@@ -161,7 +161,7 @@
     "meta": {
         "name": "Splunk_TA_dummy_data",
         "restRoot": "Splunk_TA_dummy_data",
-        "version": "5.21.0R2dc94b82",
+        "version": "5.21.0R43e797b8",
         "displayName": "Splunk_TA_dummy_data",
         "schemaVersion": "0.0.3"
     }
