+++ splunk_add_on_ucc_framework/commands/modular_alert_builder/alert_actions_conf_gen.py
@@ -23,16 +23,16 @@ from os import path as op
 
 from jinja2 import Environment, FileSystemLoader
 
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder import (
     alert_actions_exceptions as aae,
 )
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder import (
     arf_consts as ac,
 )
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core.alert_actions_helper import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_helper import (
     write_file,
 )
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core.alert_actions_merge import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_merge import (
     remove_alert_from_conf_file,
 )
