+++ splunk_add_on_ucc_framework/commands/modular_alert_builder/alert_actions_helper.py
@@ -20,7 +20,7 @@ from os import makedirs, remove
 
 import addonfactory_splunk_conf_parser_lib as conf_parser
 
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core.alert_actions_merge import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_merge import (
     merge_conf_file,
 )
