+++ splunk_add_on_ucc_framework/commands/modular_alert_builder/alert_actions_html_gen.py
@@ -21,16 +21,16 @@ from re import search
 
 from jinja2 import Environment, FileSystemLoader
 
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder import (
     alert_actions_exceptions as aae,
 )
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder import (
     arf_consts as ac,
 )
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core.alert_actions_helper import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_helper import (
     write_file,
 )
-from splunk_add_on_ucc_framework.commands.modular_alert_builder.build_core.alert_actions_template import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_template import (
     AlertActionsTemplateMgr,
 )
