+++ splunk_add_on_ucc_framework/commands/modular_alert_builder/arf_template/eventtypes.conf.template
